# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Título de tu API",
        "description": "Descripción de tu API",
        "contact": {
            "name": "Nombre del Contacto",
            "email": "contacto@ejemplo.com"
        },
        "license": {
            "name": "Licencia",
            "url": "http://www.ejemplo.com/licencia"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/cards": {
            "get": {
                "tags": [
                    "Cartas"
                ],
                "summary": "Obtener lista de cartas",
                "operationId": "be61f7fd1c68684fbae62d87cc2d6136",
                "responses": {
                    "200": {
                        "description": "Cartas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocurrió un error al obtener las cartas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cartas"
                ],
                "summary": "Crear una nueva carta",
                "operationId": "0549fc901f18c03961ca7d4700873135",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "stroke": {
                                        "type": "integer"
                                    },
                                    "defense": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carta creada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación u otro problema"
                    }
                }
            }
        },
        "/api/cards/{id}": {
            "get": {
                "tags": [
                    "Cartas"
                ],
                "summary": "Obtener una carta por ID",
                "operationId": "79ffdbbc185a77ab1705b7404104f2c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carta obtenida exitosamente"
                    },
                    "404": {
                        "description": "Carta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cartas"
                ],
                "summary": "Actualizar una carta",
                "operationId": "3c1d28f138a08adc4eed1da6dbe4257c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "stroke": {
                                        "type": "integer"
                                    },
                                    "defense": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carta actualizada exitosamente"
                    },
                    "400": {
                        "description": "Error al actualizar la carta"
                    }
                }
            }
        },
...
    },
...
}